requirements.txt
streamlit
google-generativeai
graphviz
python-dotenv